pyproject.toml
[project]
name = "ima-ink-editor"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "marimo>=0.10",
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9",
]
